requirements/base.txt
django==4.2
pillow==10.4.0
django-environ==0.11.2
django-tinymce==4.1.0
django-hitcount==1.3.5
xlwt==1.3.0